requirements.txt
fastapi
uvicorn
python-dotenv
cryptography
pydantic
sqlalchemy
psycopg2-binary
passlib
email-validator
passlib[bcrypt]
python-jose[cryptography]
sqlmodel
pyjwt
bcrypt
pytest
flake8
alembic
celery
redis